refactor(notification): correct names for layers
