+++ services/notification/src/notification/email-templates/mail.templates.ts
