+++ services/notification/src/notification/notification.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { MailModule } from 'src/mail/mail.module';
-import { NotificationService } from './application/subscription-notification.service';
+import { NotificationService } from './application/notification.service';
 import { NotificationController } from './presentation/controllers/notification.controller';
 
 @Module({
