+++ services/notification/src/notification/presentation/dto/weather-data.type.ts
